Bandage fixes for the CreateGroupActivity flow.
Fixes a bug where the Group would be saved before the facebookGroupId was set, which resulted in a null value exception. Two things to note. Post-merging, Find Matches will always return the current list of groups until we figure out a way to asynchronously fetch matching groups as pass them to the relevant fragment. Second, we might want to consider revising the flow of CreateGroupActivity since we're currently (1) Setting our callback for FB group creation, (2) Doing group creation stuff, (3) Saving from the group creation stuff, (4) Calling for the FB group creation to happen, but all of this is disparately organized within CreateGroupActivity.